// src/types/UserConfig.ts
// Code generated by schema2code at 2026-03-07 19:34:28. DO NOT EDIT.


import { SummarizationConfig } from './SummarizationConfig';

import { MemoryConfig } from './MemoryConfig';

import { ModelProfileConfig } from './ModelProfileConfig';

import { ImageGenerationConfig } from './ImageGenerationConfig';

import { CircuitBreakerConfig } from './CircuitBreakerConfig';

import { GPUConfig } from './GpuConfig';

import { ParameterOptimizationConfig } from './ParameterOptimizationConfig';

import { WorkflowConfig } from './WorkflowConfig';

import { ToolConfig } from './ToolConfig';

import { EventStreamConfig } from './EventStreamConfig';

import { PreferencesConfig } from './PreferencesConfig';



/**
 * User-specific configuration
 */
export interface UserConfig {
  /**
   * User ID
   */
  user_id: string;
  summarization?: SummarizationConfig;
  memory?: MemoryConfig;
  model_profiles?: ModelProfileConfig;
  image_generation?: ImageGenerationConfig;
  circuit_breaker: CircuitBreakerConfig;
  gpu_config: GPUConfig;
  parameter_optimization: ParameterOptimizationConfig;
  workflow: WorkflowConfig;
  tool?: ToolConfig;
  event_stream?: EventStreamConfig;
  preferences?: PreferencesConfig;
}