// src/types/ParameterOptimizationConfig.ts
// Code generated by schema2code at 2026-03-07 19:34:28. DO NOT EDIT.


import { PerformanceParameter } from './PerformanceParameter';

import { CrashPrevention } from './CrashPrevention';



/**
 * Configuration for optimizing LLM initialization parameters to find maximum values
 */
export interface ParameterOptimizationConfig {
  /**
   * Enable parameter optimization to find maximum viable values
   */
  enabled: boolean;
  /**
   * List of performance parameters to optimize
   */
  parameters: (PerformanceParameter)[];
  /**
   * Settings to prevent container crashes during optimization
   */
  crash_prevention: CrashPrevention;
}