// src/types/ParameterTuningStrategy.ts
// Code generated by schema2code at 2026-03-07 19:34:28. DO NOT EDIT.


/**
 * Strategy for tuning parameters optimally.
 */
export type ParameterTuningStrategy = 'binary_search' | 'exponential_backoff' | 'conservative_increment';

/**
 * Constant values for ParameterTuningStrategy
 */
export const ParameterTuningStrategyValues = {
  /** binary_search */
  BINARY_SEARCH: 'binary_search',
  /** exponential_backoff */
  EXPONENTIAL_BACKOFF: 'exponential_backoff',
  /** conservative_increment */
  CONSERVATIVE_INCREMENT: 'conservative_increment'
} as const;